requirements.txt
python-telegram-bot==11.1.0
requests~=2.28.1
python-dotenv~=0.21.0
vk-api==11.9.9
redis~=4.5.4